stable-diffusion/requirements.txt
accelerate==0.23.0
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
comm==0.1.4
controlnet-aux==0.0.7
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
diffusers==0.22.0.dev0
dill==0.3.7
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
huggingface-hub==0.17.2
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.11.2
sentry-sdk==1.31.0
setproctitle==1.3.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tifffile==2023.9.18
timm==0.9.7
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.1.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.2
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.10
wcwidth==0.2.6
Werkzeug==2.3.7
wheel==0.38.4
xformers==0.0.21
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0